requirements.txt
backports.functools-lru-cache==1.5
backports.weakref==1.0.post1
bleach==1.5.0
brewer2mpl==1.4.1
certifi==2018.8.24
cycler==0.10.0
decorator==4.3.0
dicom==0.9.9
funcsigs==1.0.2
futures==3.1.1
ggplot==0.11.2
html5lib==0.9999999
jmespath==0.9.3
kiwisolver==1.0.1
Markdown==2.6.11
matplotlib
mock==2.0.0
networkx==2.1
numpy==1.13.3
patsy==0.5.0
pbr==4.2.0
Pillow==5.2.0
PIMS==0.4.1
protobuf==3.6.1
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2018.5
PyWavelets
PyYAML==3.13
rsa==3.4.2
s3transfer==0.1.10
scikit-image==0.13.0
scikit-learn==0.19.1
scipy==1.1.0
six==1.11.0
slicerator==0.9.8
statsmodels==0.9.0
subprocess32==3.5.2
tensorflow-tensorboard
trackpy==0.3.1
Werkzeug==0.14.1
